requirements/common.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
PyJWT==1.7.1
sqlparse==0.5.1
tzdata==2024.1
